convert test dirs to packages (#1204)
* convert test dirs to packages

* fix cross module import of fixture in test_location

* add whats new
