+++ pvlib/tests/iotools/test_surfrad.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pytest
 
 from pvlib.iotools import surfrad
-from conftest import DATA_DIR, RERUNS, RERUNS_DELAY
+from ..conftest import DATA_DIR, RERUNS, RERUNS_DELAY
 
 testfile = DATA_DIR / 'surfrad-slv16001.dat'
 network_testfile = ('ftp://aftp.cmdl.noaa.gov/data/radiation/surfrad/'
