+++ pvlib/tests/test_modelchain.py
@@ -10,10 +10,10 @@ from pvlib.tracking import SingleAxisTracker
 from pvlib.location import Location
 from pvlib._deprecation import pvlibDeprecationWarning
 
-from conftest import assert_series_equal, assert_frame_equal
+from .conftest import assert_series_equal, assert_frame_equal
 import pytest
 
-from conftest import fail_on_pvlib_version, requires_tables
+from .conftest import fail_on_pvlib_version, requires_tables
 
 
 @pytest.fixture(scope='function')
