+++ pvlib/tests/test_solarposition.py
@@ -5,14 +5,14 @@ import warnings
 import numpy as np
 import pandas as pd
 
-from conftest import assert_frame_equal, assert_series_equal
+from .conftest import assert_frame_equal, assert_series_equal
 from numpy.testing import assert_allclose
 import pytest
 
 from pvlib.location import Location
 from pvlib import solarposition, spa
 
-from conftest import requires_ephem, requires_spa_c, requires_numba
+from .conftest import requires_ephem, requires_spa_c, requires_numba
 
 
 # setup times and locations to be tested.
@@ -23,31 +23,6 @@ tus = Location(32.2, -111, 'US/Arizona', 700)  # no DST issues possible
 times_localized = times.tz_localize(tus.tz)
 
 tol = 5
-
-
-@pytest.fixture()
-def golden():
-    return Location(39.742476, -105.1786, 'America/Denver', 1830.14)
-
-
-@pytest.fixture()
-def golden_mst():
-    return Location(39.742476, -105.1786, 'MST', 1830.14)
-
-
-@pytest.fixture()
-def expected_solpos():
-    return _expected_solpos_df()
-
-
-# hack to make tests work without too much modification while avoiding
-# pytest 4.0 inability to call features directly
-def _expected_solpos_df():
-    return pd.DataFrame({'elevation': 39.872046,
-                         'apparent_zenith': 50.111622,
-                         'azimuth': 194.340241,
-                         'apparent_elevation': 39.888378},
-                        index=['2003-10-17T12:30:30Z'])
 
 
 @pytest.fixture()
@@ -455,7 +430,7 @@ def test_get_solarposition_error(golden):
 
 
 @pytest.mark.parametrize("pressure, expected", [
-    (82000, _expected_solpos_df()),
+    (82000, 'expected_solpos'),
     (90000, pd.DataFrame(
         np.array([[39.88997,   50.11003,  194.34024,   39.87205,   14.64151,
                    50.12795]]),
@@ -463,13 +438,16 @@ def test_get_solarposition_error(golden):
                  'elevation', 'equation_of_time', 'zenith'],
         index=['2003-10-17T12:30:30Z']))
     ])
-def test_get_solarposition_pressure(pressure, expected, golden):
+def test_get_solarposition_pressure(
+        pressure, expected, golden, expected_solpos):
     times = pd.date_range(datetime.datetime(2003, 10, 17, 13, 30, 30),
                           periods=1, freq='D', tz=golden.tz)
     ephem_data = solarposition.get_solarposition(times, golden.latitude,
                                                  golden.longitude,
                                                  pressure=pressure,
                                                  temperature=11)
+    if isinstance(expected, str) and expected == 'expected_solpos':
+        expected = expected_solpos
     this_expected = expected.copy()
     this_expected.index = times
     this_expected = np.round(this_expected, 5)
@@ -478,7 +456,7 @@ def test_get_solarposition_pressure(pressure, expected, golden):
 
 
 @pytest.mark.parametrize("altitude, expected", [
-    (1830.14, _expected_solpos_df()),
+    (1830.14, 'expected_solpos'),
     (2000, pd.DataFrame(
         np.array([[39.88788,   50.11212,  194.34024,   39.87205,   14.64151,
                    50.12795]]),
@@ -486,13 +464,16 @@ def test_get_solarposition_pressure(pressure, expected, golden):
                  'elevation', 'equation_of_time', 'zenith'],
         index=['2003-10-17T12:30:30Z']))
     ])
-def test_get_solarposition_altitude(altitude, expected, golden):
+def test_get_solarposition_altitude(
+        altitude, expected, golden, expected_solpos):
     times = pd.date_range(datetime.datetime(2003, 10, 17, 13, 30, 30),
                           periods=1, freq='D', tz=golden.tz)
     ephem_data = solarposition.get_solarposition(times, golden.latitude,
                                                  golden.longitude,
                                                  altitude=altitude,
                                                  temperature=11)
+    if isinstance(expected, str) and expected == 'expected_solpos':
+        expected = expected_solpos
     this_expected = expected.copy()
     this_expected.index = times
     this_expected = np.round(this_expected, 5)
